PR comment refactoring
