+++ app/src/main/java/com/hedvig/app/feature/offer/ui/faq/FAQItem.kt
@@ -1,0 +1,6 @@
+package com.hedvig.app.feature.offer.ui.faq
+
+data class FAQItem(
+    val headline: String,
+    val body: String,
+)
